Decoupled flow rule cache from flow rule manager
Heavy refactoring of all the relevant components
