+++ configure.in
@@ -1348,9 +1348,10 @@ fi
 
 
 AC_ARG_ENABLE([flow-api], [AS_HELP_STRING([--enable-flow-api], [Support for the HW flow classification API. Also needs DPDK>=20.02 ])], [:], [enable_flow_api=no])
-AC_CHECK_FILE($srcdir/include/click/flowparser.hh, [ac_have_flow_parser_h=yes], [ac_have_flow_parser_h=no])
-AC_CHECK_FILE($srcdir/include/click/flowparserglue.hh, [ac_have_flow_parser_g_h=yes], [ac_have_flow_parser_g_h=no])
-if test "x$ac_have_flow_parser_h$ac_have_flow_parser_g_h$RTE_VER_YEAR$enable_flow_api" = "xyesyes20yes"; then
+AC_CHECK_FILE($srcdir/include/click/flowruleparser.hh, [ac_have_flow_rule_parser_h=yes], [ac_have_flow_rule_parser_h=no])
+AC_CHECK_FILE($srcdir/include/click/flowrulecache.hh, [ac_have_flow_rule_cache_h=yes], [ac_have_flow_rule_cache_h=no])
+AC_CHECK_FILE($srcdir/include/click/flowrulemanager.hh, [ac_have_flow_rule_man_h=yes], [ac_have_flow_rule_man_h=no])
+if test "x$ac_have_flow_rule_parser_h$ac_have_flow_rule_cache_h$ac_have_flow_rule_man_h$RTE_VER_YEAR$enable_flow_api" = "xyesyesyes20yes"; then
     have_flow_api=yes
     AC_DEFINE([HAVE_FLOW_API])
 else
@@ -2556,7 +2557,7 @@ AC_OUTPUT
 feedback="no"
 
 if test "$enable_flow_api$have_flow_api" = "yesno"; then
-    feedback="To enable DPDK's Flow Parser (rte_flow API), use a DPDK version >= 20.02."
+    feedback="To enable DPDK's Flow Rule Manager (rte_flow API), use a DPDK version >= 20.02."
 fi
 
 echo "
