+++ elements/userlevel/fromdpdkdevice.cc
@@ -34,7 +34,7 @@
 #endif
 
 #if RTE_VERSION >= RTE_VERSION_NUM(20,2,0,0)
-    #include <click/flowparser.hh>
+    #include <click/flowrulemanager.hh>
 #endif
 
 CLICK_DECLS
@@ -165,9 +165,9 @@ int FromDPDKDevice::configure(Vector<String> &conf, ErrorHandler *errh)
         _dev->set_init_rss_max(max_rss);
 
 #if RTE_VERSION >= RTE_VERSION_NUM(20,2,0,0)
-    if ((mode == FlowParser::DISPATCHING_MODE) && (flow_rules_filename.empty())) {
+    if ((mode == FlowRuleManager::DISPATCHING_MODE) && (flow_rules_filename.empty())) {
         errh->warning(
-            "DPDK Flow Parser (port %s): FLOW_RULES_FILE is not set, "
+            "DPDK Flow Rule Manager (port %s): FLOW_RULES_FILE is not set, "
             "hence this NIC can only be configured by the handlers",
             dev.c_str()
         );
@@ -510,31 +510,31 @@ String FromDPDKDevice::statistics_handler(Element *e, void *thunk)
     #if RTE_VERSION >= RTE_VERSION_NUM(20,2,0,0)
         case h_rules_list: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return FlowParser::get_flow_parser(port_id)->flow_rules_list();
+            return FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rules_list();
         }
         case h_rules_list_with_hits: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return FlowParser::get_flow_parser(port_id)->flow_rules_list(true);
+            return FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rules_list(true);
         }
         case h_rules_ids_global: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return FlowParser::get_flow_parser(port_id)->flow_rule_ids_global();
+            return FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rule_ids_global();
         }
         case h_rules_ids_internal: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return FlowParser::get_flow_parser(port_id)->flow_rule_ids_internal();
+            return FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rule_ids_internal();
         }
         case h_rules_count: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return String(FlowParser::get_flow_parser(port_id)->flow_rules_count_explicit());
+            return String(FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rules_count_explicit());
         }
         case h_rules_count_with_hits: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return String(FlowParser::get_flow_parser(port_id)->flow_rules_with_hits_count());
+            return String(FlowRuleManager::get_flow_rule_mgr(port_id)->flow_rules_with_hits_count());
         }
         case h_rules_isolate: {
             portid_t port_id = fd->get_device()->get_port_id();
-            return String(FlowParser::isolated(port_id) ? "1" : "0");
+            return String(FlowRuleManager::isolated(port_id) ? "1" : "0");
         }
     #endif
         case h_nombufs:
@@ -623,8 +623,8 @@ int FromDPDKDevice::flow_handler(
     }
 
     portid_t port_id = fd->get_device()->get_port_id();
-    FlowParser *flow_parser = FlowParser::get_flow_parser(port_id, errh);
-    assert(flow_parser);
+    FlowRuleManager *flow_rule_mgr = FlowRuleManager::get_flow_rule_mgr(port_id, errh);
+    assert(flow_rule_mgr);
 
     switch((uintptr_t) thunk) {
         case h_rule_add: {
@@ -638,18 +638,18 @@ int FromDPDKDevice::flow_handler(
             }
 
             // Detect and remove unwanted components
-            if (!FlowParser::filter_rule(rule)) {
-                return errh->error("DPDK Flow Parser (port %u): Invalid rule '%s'", port_id, rule.c_str());
+            if (!FlowRuleManager::flow_rule_filter(rule)) {
+                return errh->error("DPDK Flow Rule Manager (port %u): Invalid rule '%s'", port_id, rule.c_str());
             }
 
             rule = "flow create " + String(port_id) + " " + rule;
 
             // Parse the queue index to infer the CPU core
-            String queue_index_str = FlowParser::fetch_token_after_keyword((char *) rule.c_str(), "queue index");
+            String queue_index_str = FlowRuleManager::fetch_token_after_keyword((char *) rule.c_str(), "queue index");
             int core_id = atoi(queue_index_str.c_str());
 
-            const uint32_t int_rule_id = flow_parser->get_flow_cache()->next_internal_rule_id();
-            if (flow_parser->flow_rule_install(int_rule_id, (long) int_rule_id, core_id, rule) != 0) {
+            const uint32_t int_rule_id = flow_rule_mgr->flow_rule_cache()->next_internal_rule_id();
+            if (flow_rule_mgr->flow_rule_install(int_rule_id, (long) int_rule_id, core_id, rule) != 0) {
                 return -1;
             }
 
@@ -676,18 +676,18 @@ int FromDPDKDevice::flow_handler(
             }
 
             // Batch deletion
-            return flow_parser->flow_rules_delete((uint32_t *) rule_ids, rules_nb);
+            return flow_rule_mgr->flow_rules_delete((uint32_t *) rule_ids, rules_nb);
         }
         case h_rules_isolate: {
             if (input.empty()) {
-                return errh->error("DPDK Flow Parser (port %u): Specify isolation mode (true/1 -> isolation, otherwise no isolation)", port_id);
+                return errh->error("DPDK Flow Rule Manager (port %u): Specify isolation mode (true/1 -> isolation, otherwise no isolation)", port_id);
             }
             bool status = (input.lower() == "true") || (input.lower() == "1") ? true : false;
-            FlowParser::set_isolation_mode(port_id, status);
+            FlowRuleManager::set_isolation_mode(port_id, status);
             return 0;
         }
         case h_rules_flush: {
-            return flow_parser->flow_rules_flush();
+            return flow_rule_mgr->flow_rules_flush();
         }
     }
 
@@ -758,15 +758,15 @@ int FromDPDKDevice::xstats_handler(
         case h_rule_byte_count:
         case h_rule_packet_hits: {
             portid_t port_id = fd->get_device()->get_port_id();
-            FlowParser *flow_parser = FlowParser::get_flow_parser(port_id, errh);
-            assert(flow_parser);
+            FlowRuleManager *flow_rule_mgr = FlowRuleManager::get_flow_rule_mgr(port_id, errh);
+            assert(flow_rule_mgr);
             if (input == "") {
                 return errh->error("Aggregate flow rule counters are not supported. Please specify a rule number to query");
             } else {
                 const uint32_t rule_id = atoi(input.c_str());
                 int64_t matched_pkts = -1;
                 int64_t matched_bytes = -1;
-                flow_parser->flow_rule_query(rule_id, matched_pkts, matched_bytes);
+                flow_rule_mgr->flow_rule_query(rule_id, matched_pkts, matched_bytes);
                 if (op == (int) h_rule_packet_hits) {
                     input = String(matched_pkts);
                 } else {
@@ -777,9 +777,9 @@ int FromDPDKDevice::xstats_handler(
         }
         case h_rules_aggr_stats: {
             portid_t port_id = fd->get_device()->get_port_id();
-            FlowParser *flow_parser = FlowParser::get_flow_parser(port_id, errh);
-            assert(flow_parser);
-            input = flow_parser->flow_rule_aggregate_stats();
+            FlowRuleManager *flow_rule_mgr = FlowRuleManager::get_flow_rule_mgr(port_id, errh);
+            assert(flow_rule_mgr);
+            input = flow_rule_mgr->flow_rule_aggregate_stats();
             return 0;
         }
     #endif
@@ -822,9 +822,9 @@ void FromDPDKDevice::add_handlers()
     set_handler("queue_packets", Handler::f_read | Handler::f_read_param, xstats_handler, h_stats_packets);
     set_handler("queue_bytes", Handler::f_read | Handler::f_read_param, xstats_handler, h_stats_bytes);
 #if RTE_VERSION >= RTE_VERSION_NUM(20,2,0,0)
-    set_handler(FlowParser::FLOW_RULE_PACKET_HITS, Handler::f_read | Handler::f_read_param, xstats_handler, h_rule_packet_hits);
-    set_handler(FlowParser::FLOW_RULE_BYTE_COUNT,  Handler::f_read | Handler::f_read_param, xstats_handler, h_rule_byte_count);
-    set_handler(FlowParser::FLOW_RULE_AGGR_STATS,  Handler::f_read | Handler::f_read_param, xstats_handler, h_rules_aggr_stats);
+    set_handler(FlowRuleManager::FLOW_RULE_PACKET_HITS, Handler::f_read | Handler::f_read_param, xstats_handler, h_rule_packet_hits);
+    set_handler(FlowRuleManager::FLOW_RULE_BYTE_COUNT,  Handler::f_read | Handler::f_read_param, xstats_handler, h_rule_byte_count);
+    set_handler(FlowRuleManager::FLOW_RULE_AGGR_STATS,  Handler::f_read | Handler::f_read_param, xstats_handler, h_rules_aggr_stats);
 #endif
 
     add_read_handler("active", read_handler, h_active);
@@ -854,17 +854,17 @@ void FromDPDKDevice::add_handlers()
     add_read_handler("nombufs",statistics_handler, h_nombufs);
 
 #if RTE_VERSION >= RTE_VERSION_NUM(20,2,0,0)
-    add_write_handler(FlowParser::FLOW_RULE_ADD,     flow_handler, h_rule_add,    0);
-    add_write_handler(FlowParser::FLOW_RULE_DEL,     flow_handler, h_rules_del,   0);
-    add_write_handler(FlowParser::FLOW_RULE_ISOLATE, flow_handler, h_rules_isolate, 0);
-    add_write_handler(FlowParser::FLOW_RULE_FLUSH,   flow_handler, h_rules_flush, 0);
-    add_read_handler (FlowParser::FLOW_RULE_IDS_GLB,         statistics_handler, h_rules_ids_global);
-    add_read_handler (FlowParser::FLOW_RULE_IDS_INT,         statistics_handler, h_rules_ids_internal);
-    add_read_handler (FlowParser::FLOW_RULE_LIST,            statistics_handler, h_rules_list);
-    add_read_handler (FlowParser::FLOW_RULE_LIST_WITH_HITS,  statistics_handler, h_rules_list_with_hits);
-    add_read_handler (FlowParser::FLOW_RULE_COUNT,           statistics_handler, h_rules_count);
-    add_read_handler (FlowParser::FLOW_RULE_COUNT_WITH_HITS, statistics_handler, h_rules_count_with_hits);
-    add_read_handler (FlowParser::FLOW_RULE_ISOLATE,         statistics_handler, h_rules_isolate);
+    add_write_handler(FlowRuleManager::FLOW_RULE_ADD,     flow_handler, h_rule_add,    0);
+    add_write_handler(FlowRuleManager::FLOW_RULE_DEL,     flow_handler, h_rules_del,   0);
+    add_write_handler(FlowRuleManager::FLOW_RULE_ISOLATE, flow_handler, h_rules_isolate, 0);
+    add_write_handler(FlowRuleManager::FLOW_RULE_FLUSH,   flow_handler, h_rules_flush, 0);
+    add_read_handler (FlowRuleManager::FLOW_RULE_IDS_GLB,         statistics_handler, h_rules_ids_global);
+    add_read_handler (FlowRuleManager::FLOW_RULE_IDS_INT,         statistics_handler, h_rules_ids_internal);
+    add_read_handler (FlowRuleManager::FLOW_RULE_LIST,            statistics_handler, h_rules_list);
+    add_read_handler (FlowRuleManager::FLOW_RULE_LIST_WITH_HITS,  statistics_handler, h_rules_list_with_hits);
+    add_read_handler (FlowRuleManager::FLOW_RULE_COUNT,           statistics_handler, h_rules_count);
+    add_read_handler (FlowRuleManager::FLOW_RULE_COUNT_WITH_HITS, statistics_handler, h_rules_count_with_hits);
+    add_read_handler (FlowRuleManager::FLOW_RULE_ISOLATE,         statistics_handler, h_rules_isolate);
 #endif
 
     add_read_handler("mtu",read_handler, h_mtu);
